project/requirements.txt
Django
pandas==1.5.3
psycopg2-binary==2.9.1
gunicorn==20.1.0
celery==5.3.6
cryptography==38.0.4
Scrapy==2.11.0
scrapy-djangoitem==1.1.1
pyOpenSSL==22.0.0
redis==5.0.3
djangorestframework
django-cors-headers
django-filter
requests
paho-mqtt==1.6.1
pytz
#numpy==1.19.5
#python-dateutil==2.8.2